+++ source/sprites/sprite_utils.h
@@ -73,7 +73,7 @@ namespace sprite_utils {
     }
 
     //!> sets xy for given sprite entry
-    void set_entry_xy(SpriteInfo *i, u16 x, u16 y);
+    void set_entry_xy(SpriteInfo *i, int x, int y);
 
     u8 *get_frame(u8 *tileset, int size, int frame_offset);
 }
